+++ flink-examples/flink-examples-build-helper/pom.xml
@@ -35,7 +35,7 @@ under the License.
 	<modules>
 		<module>flink-examples-streaming-twitter</module>
 		<module>flink-examples-streaming-state-machine</module>
-		<module>flink-examples-streaming-pubsub</module>
+		<module>flink-examples-streaming-gcp-pubsub</module>
 	</modules>
 
 </project>
